+++ bin/blocks/carousel.sh
@@ -2,8 +2,7 @@
 # @Author: Roni Laukkarinen
 # @Date:   2022-02-10 10:44:02
 # @Last Modified by:   Roni Laukkarinen
-# @Last Modified time: 2022-02-10 15:47:32
-
+# @Last Modified time: 2022-02-11 15:33:54
 
 # // New files/Dependencies (this file will install them)::
 # // ├── sass/gutenberg/blocks/_carousel.scss (automatic from get-block.sh)