+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### [Unreleased]
+
+* New feature: You can now add new empty blocks!
+
 ### 1.0.4: 2021-02-10
 
 * Fix 5.9 related WSOD
@@ -6,7 +10,6 @@
 * newblock.sh block importer: Add accordion
 * newblock.sh block importer: Add carousel
 * New feature: Activates the block ACF fields automatically (no extra step with "Sync available" > Import)
-* New feature: You can now add new empty blocks!
 
 ### 1.0.3: 2021-02-08